Fix issue with YouTube videos not playing
YouTube had implemented a new transport protocol. This was not supported by the 0.8 revision of the Youtube_dl this was fixed by updating the crate to version 0.9.
The new revision of the crate had increased safety by wrapping the result in an option. The source has been modified to extract the data to allow it to work with this new revision.